:truck: Move BarChartSelect to filter module
