+++ src/frontend/modules/filter/components/BarChartSelect.tsx
